Relocate helpers.js
